refactor(M6 PR2): extract the help overlay into controller/help.rs
Move the help feature's impl Controller methods (open/close, key + mouse
handling, scroll, the HelpView projection, help_state) out of the god-module
into a cohesive submodule. Pure relocation: methods called from the parent
dispatch are pub(super); intra-module helpers stay private. Behavior-preserving
(use super::* keeps them the same Controller). mod.rs: 3256 -> ~3022 lines.
